--- src/templates/blog-post.css
@@ -1,7 +1,0 @@
-.postByline, .backLink {
-    font-size: 0.75rem;
-}
-
-.postBody {
-    margin-top: 2rem;
-}
